XEP-0060: Support 'integer-or-max' data type
Support the 'integer-or-max' data type, which allows for requesting the
maximum limit for certain node configuration options by specifying 'max'
instead of an integer.  This was added to XEP-0060, revision 1.17.0 (and
clarified in revision 1.20.0) for the pubsub#children_max,
pubsub#item_expire, and pubsub#max_items options.

Thanks to Ammonit Measurement GmbH for sponsoring this work.