specs: mix: Fix mix:core:1 mix_participant and drop old version
The mix_participant record is only used for the participants node of a
mix channel. This node cannot contain both namespaces at the same time,
thus this commit drops the old version and it also fixes the definition
of the element in mix:core:1. The attributes have been replaced with xml
elements for nick and jid.